feat: Refactor shopping cart and order item models to use int for quantity and update related DTOs

+++ order-microservice/src/main/java/com/metao/book/order/application/cart/ShoppingCartItem.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 @Builder
 public record ShoppingCartItem(
     String sku,
-    BigDecimal quantity,
+    int quantity,
     BigDecimal price,
     Currency currency
 ) {
